requirements.txt
PySide6-Essentials>=6.7.0
